requirements.txt
newspaper4k
nltk
pandas
streamlit
sqlalchemy